+++ lib/src/interceptor/cookie_response_interceptor.dart
@@ -1,5 +1,7 @@
 import 'dart:io';
 
+import 'package:dart_console/dart_console.dart';
+import 'package:logging/logging.dart';
 import 'package:reserve/reserve.dart';
 
 class CookieResponseInterceptor extends ResponseInterceptor {
@@ -29,8 +31,16 @@ class CookieResponseInterceptor extends ResponseInterceptor {
     ReServeResponse response,
   ) {
     final headers = response.headers.all.where((h) => h.key == _kHeaderName);
+    if (headers.isEmpty) {
+      return response;
+    }
 
     final cookies = <ReServeCookie>[];
+
+    logger.finest('Modified cookies for response:');
+    String maxln(String input, {int length = 16}) =>
+        input.length < length ? input : '${input.substring(0, length - 3)}...';
+    final cookieTable = <List<String>>[];
 
     for (final header in headers) {
       final cookie = ReServeCookie.fromCookie(
@@ -39,12 +49,47 @@ class CookieResponseInterceptor extends ResponseInterceptor {
 
       final host = config.origin?.host ?? config.host;
 
-      cookies.add(
-        cookie.copyWith(
-          domain: cookie.domain == null ? null : '.$host',
-          secure: allowSecure ? cookie.secure : false,
-        ),
+      final copied = cookie.copyWith(
+        domain: '.$host',
+        secure: allowSecure ? cookie.secure : false,
       );
+
+      if (logger.isLoggable(Level.FINEST)) {
+        cookieTable.add([
+          maxln(copied.name),
+          maxln(copied.value),
+          maxln(copied.domain ?? ''),
+          maxln(copied.path ?? '/'),
+          maxln(copied.httpOnly.toString()),
+          maxln(copied.secure.toString()),
+          maxln(copied.maxAge.toString()),
+          maxln(copied.expires?.toString() ?? '', length: 20),
+        ]);
+      }
+      cookies.add(copied);
+    }
+
+    if (logger.isLoggable(Level.FINEST)) {
+      cookieTable.sort(
+        (a, b) => a[0].toLowerCase().compareTo(b[0].toLowerCase()),
+      );
+      final table = Table()
+        ..insertColumn(header: 'Name')
+        ..insertColumn(header: 'Value')
+        ..insertColumn(header: 'Domain')
+        ..insertColumn(header: 'Path')
+        ..insertColumn(header: 'HttpOnly')
+        ..insertColumn(header: 'Secure')
+        ..insertColumn(header: 'MaxAge')
+        ..insertColumn(header: 'Expires')
+        ..insertRows(cookieTable);
+
+      // The table is already long and the previous line has the timestamp and
+      // logger info.  In this case, directly write the output rather than going
+      // through the logger.
+      //
+      // ignore: avoid_print
+      print(table);
     }
 
     final result = List<ReServeHeader>.from(response.headers.all)