+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0+1
+## 1.0.1
+
+- Added logging for cookie processing
+- Added `embed` to be able to print version information
+- Added ability to specific the root key to find the configuration at
+- Added `web_dev_config.yaml` as a search path
+
+## 1.0.0
 
 - Initial Beta Release: API stable but documentation and tests in flight